Organise components per atomic design
